+++ submission/src/epilot/sdkconfiguration.py
@@ -15,14 +15,13 @@ SERVERS = [
 @dataclass
 class SDKConfiguration:
     client: requests.Session
-    security_client: requests.Session
     server_url: str = ''
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '2.0.0'
-    gen_version: str = '2.181.1'
-    user_agent: str = 'speakeasy-sdk/python 2.0.0 2.181.1 1.0.0 epilot-submission'
+    sdk_version: str = '2.0.1'
+    gen_version: str = '2.185.0'
+    user_agent: str = 'speakeasy-sdk/python 2.0.1 2.185.0 1.0.0 epilot-submission'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: